Cleanup runtime scripts
