+++ includes/runtime/src/hooks/use-bound-group-extractor.js
@@ -1,9 +1,23 @@
 import { useEffect } from '@wordpress/element';
-import { registerPlugin } from '@wordpress/plugins';
 import { useEntityProp } from '@wordpress/core-data';
 import { useSelect } from '@wordpress/data';
 import { store as editorStore } from '@wordpress/block-editor';
 import { serialize } from '@wordpress/blocks';
+import { POST_TYPE } from '../constants';
+
+export const useBoundGroupExtractor = () => {
+	const blocks = useSelect( ( select ) => select( editorStore ).getBlocks() );
+
+	const [ , setMeta ] = useEntityProp( 'postType', POST_TYPE, 'meta' );
+
+	useEffect( () => {
+		const boundBlocks = findBoundBlocks( blocks );
+
+		setMeta( boundBlocks );
+	}, [ blocks, setMeta ] );
+
+	return null;
+};
 
 const findBoundBlocks = ( blocks, acc = {} ) => {
 	for ( const block of blocks ) {
@@ -24,25 +38,3 @@ const findBoundBlocks = ( blocks, acc = {} ) => {
 
 	return acc;
 };
-
-const CreateContentModelBoundGroupExtractor = () => {
-	const blocks = useSelect( ( select ) => select( editorStore ).getBlocks() );
-
-	const [ , setMeta ] = useEntityProp(
-		'postType',
-		window.contentModelFields.postType,
-		'meta'
-	);
-
-	useEffect( () => {
-		const boundBlocks = findBoundBlocks( blocks );
-
-		setMeta( boundBlocks );
-	}, [ blocks, setMeta ] );
-
-	return null;
-};
-
-registerPlugin( 'create-content-model-bound-group-extractor', {
-	render: CreateContentModelBoundGroupExtractor,
-} );
