+++ includes/runtime/src/register-content-locking.js
@@ -1,0 +1,11 @@
+import { registerPlugin } from '@wordpress/plugins';
+import { useContentLocking } from './hooks/use-content-locking';
+
+export const registerContentLocking = () => {
+	registerPlugin( 'create-content-model-content-locking', {
+		render: () => {
+			// eslint-disable-next-line react-hooks/rules-of-hooks
+			useContentLocking();
+		},
+	} );
+};
